backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pocketbase>=0.15.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.4",
]
